# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online Learning Platform API",
        "description": "API documentation for the Online Learning Platform",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/student/booking": {
            "get": {
                "tags": [
                    "Booking"
                ],
                "summary": "Get bookings for authenticated student",
                "operationId": "e8cbc290436cd4f563bb8c630e5abe50",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of bookings"
                    }
                }
            },
            "post": {
                "tags": [
                    "Booking"
                ],
                "summary": "Create a new booking",
                "operationId": "470d796272ea732e099bf85400e471d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "course_id": {
                                        "type": "integer"
                                    },
                                    "service_id": {
                                        "type": "integer"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "single"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking created successfully"
                    }
                }
            }
        },
...
        "/api/student/booking/pay": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Pay for a booking (card)",
                "operationId": "ca7866ad9ded3113bd76e1eee505786b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment processed or OTP required"
                    }
                }
            }
        },
        "/api/student/booking/{bookingId}": {
            "get": {
                "tags": [
                    "Booking"
                ],
                "summary": "Get booking details",
                "operationId": "7aec6c83682ef169bf0cd97f42502e4e",
                "parameters": [
                    {
                        "name": "bookingId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking details"
                    }
                }
            }
        },
        "/api/student/booking/{bookingId}/cancel": {
            "put": {
                "tags": [
                    "Booking"
                ],
                "summary": "Cancel a booking",
                "operationId": "854b36c8319caf62e09a8d48486ad5b4",
                "parameters": [
                    {
                        "name": "bookingId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking cancelled"
                    }
                }
            }
        },
...
        "/api/courses/{id}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get course details",
                "operationId": "bc57e32b05958570eae4745a05618e50",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course details"
                    }
                }
            }
        },
        "/api/payments/direct": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Create a direct payment via HyperPay",
                "operationId": "343ed295ea639d81da58f3b3cd5740a9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "HyperPay response"
                    }
                }
            }
        },
        "/api/payments/result": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "Check payment result after 3DS redirect",
                "operationId": "748028b6f900a35f9e5f6ce830fb8a4c",
                "parameters": [
                    {
                        "name": "resourcePath",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to success or failed page"
                    }
                }
            }
        },
        "/api/sessions": {
            "get": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Get all sessions for the authenticated user",
                "operationId": "0b722a0b8a07394301d29492524e35c3",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of sessions for the authenticated user"
                    }
                }
            }
        },
...
        "/api/sessions/{sessionId}": {
            "get": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Get session details",
                "operationId": "c52f6cbf354c88760b813c01770a7aef",
                "parameters": [
                    {
                        "name": "sessionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session details"
                    }
                }
            }
        },
...
        "/api/teacher/wallet/withdrawals/{id}": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Get withdrawal details",
                "operationId": "cc14c9c51abbb5e7cd8cf2b751fe952c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Withdrawal details"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Cancel a pending withdrawal",
                "operationId": "1988527c662ca305228d8b8934540fce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Withdrawal cancelled"
                    }
                }
            }
        },
        "/api/teacher/wallet/withdrawals": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "List withdrawal requests",
                "operationId": "3d4614dc101c7c26090f8b8a38f6e4b7",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of withdrawals"
                    }
                }
            }
        }
    },
...
}